Move GithHub handler to own file
